refactor: address full branch code review findings
Creature_ManaBarrier.cs:
- Move ManaBarrierResult struct here from Player_ManaBarrier (logical home
  for a type used by both Creature and Player)
- Fix trailing whitespace inside TryAbsorbWithManaBarrier

Player_ManaBarrier.cs:
- Remove ManaBarrierRatios duplicate (defined on Creature base class)
- Remove TryAbsorbWithManaBarrier implementations (inherited via virtual
  dispatch; only GetManaBarrierRatioMod differed, now correctly overrides)
- Add 'override' keyword to GetManaBarrierRatioMod
- Fix garbled encoding in tier comments

CharmAbilityRegistry.cs:
- Add RegisteredIds property (IEnumerable<int>) to expose registered IDs
  without requiring callers to hardcode a numeric range

Player_AbilityCharms.cs:
- Move ValidateAbilityCharms() here from Player_Properties (logic method
  belongs in ability charm partial, not property partial)
- Update loop to use CharmAbilityRegistry.RegisteredIds instead of
  hardcoded 'for id = 1..50'

Player_Properties.cs:
- Remove ValidateAbilityCharms (now in Player_AbilityCharms.cs)

Gem.cs:
- Replace magic number abilityId == 1 with CharmAbilityRegistry.ManaBarrierAbilityId
- Add player.SaveBiotaToDatabase(enqueueSave: true) alongside gem save to
  persist player-side charm state (HasManaBarrier etc.) immediately

Monster_Combat.cs:
- Change preHitHealth to long to avoid uint->int truncation on high-HP
  creatures in overkill calculation

WorldObject_Magic.cs:
- Replace unsafe targetCreature.Mana.Current/MaxValue access with
  GetManaShieldSuffix() which handles null Mana vital gracefully

WorldObject_Properties.cs:
- Fix encoding corruption in ILT section divider comments
